\name{NU.Learning-internal}
\alias{NU.Learning-internal}
\alias{plot.NUcluster}
\alias{plot.confirm}
\alias{plot.KSperm}
\alias{print.NUcluster}
\alias{print.ivadj}
\alias{print.lrcagg}
\alias{print.ltdagg}
\alias{print.confirm}
\alias{print.KSperm}
\title{Internal LocalControlStrategy functions.}
\description{Internal functions for the NU.Learning-package.  Plot and print
 functions require the output object from a NU.Learning function as their
 first argument.}
\author{Bob Obenchain <wizbob@att.net>}
\keyword{methods}
\keyword{internal}
