\name{print.mlme}
\alias{print.mlme}
\title{Print Summary Statistics on Local effect-size Estimates for Patients "Most-Like-Me".}
\description{Display "Most-Like-Me" Summary Statistics for LOCAL effect-size (LTD or LRC) Distributions
  of "Nearest-Neighbors" in X-confounder Space.
}
\usage{
  \method{print}{mlme}(x, \dots)
}
\arguments{
  \item{x}{An object output by mlme().}
  \item{\dots}{Other arguments passed on to print().}
}
\value{ NULL }
\author{Bob Obenchain <wizbob@att.net>}
\seealso{\code{\link{mlme.stats}},\code{\link{plot.mlme}},\code{\link{mlme}}}
\keyword{methods}
